renamed sqlx->example and added invalid type
